Create line_clipper_1_0.py
The commentary inside this Python file describes more in-depth what the script does.  Briefly, this is a function which takes a planar polygon in R^3 and a line, both given as lists of their vertices which define their faces (you can think of the line as a 1-polygon of a single 1-face and two vertices), and returns a list of line segments the parts of the line which are visible -- that is to say, they are not occluded by F, assuming that the latter is completely opaque.